// frontend/src/api/stub/models/OrderCreate.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Restaurant POS
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.5.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { OrderStatus } from './OrderStatus';
import {
    OrderStatusFromJSON,
    OrderStatusFromJSONTyped,
    OrderStatusToJSON,
    OrderStatusToJSONTyped,
} from './OrderStatus';

/**
 * Schema for creating a new order
 * @export
 * @interface OrderCreate
 */
export interface OrderCreate {
    /**
     * 
     * @type {number}
     * @memberof OrderCreate
     */
    tableId: number;
    /**
     * 
     * @type {number}
     * @memberof OrderCreate
     */
    userId?: number | null;
    /**
     * 
     * @type {OrderStatus}
     * @memberof OrderCreate
     */
    status?: OrderStatus;
}



/**
 * Check if a given object implements the OrderCreate interface.
 */
export function instanceOfOrderCreate(value: object): value is OrderCreate {
    if (!('tableId' in value) || value['tableId'] === undefined) return false;
    return true;
}

export function OrderCreateFromJSON(json: any): OrderCreate {
    return OrderCreateFromJSONTyped(json, false);
}

export function OrderCreateFromJSONTyped(json: any, ignoreDiscriminator: boolean): OrderCreate {
    if (json == null) {
        return json;
    }
    return {
        
        'tableId': json['table_id'],
        'userId': json['user_id'] == null ? undefined : json['user_id'],
        'status': json['status'] == null ? undefined : OrderStatusFromJSON(json['status']),
    };
}

export function OrderCreateToJSON(json: any): OrderCreate {
    return OrderCreateToJSONTyped(json, false);
}

export function OrderCreateToJSONTyped(value?: OrderCreate | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'table_id': value['tableId'],
        'user_id': value['userId'],
        'status': OrderStatusToJSON(value['status']),
    };
}